# Attempt to exit early from _init if a job manifest exists and is valid.