## Description
Many methods/properties of `Job` invoke a call to `job.init()`, including the first access to `job.document` and all accesses to `job.stores` and `job.data`.

This PR optimizes `job.init()` for jobs that are already initialized, by exiting early.

## Motivation and Context
This is a performance enhancement for many common data access operations in signac.

Performance results for the script below:

| Test           | `master` | This branch (b4917d8) | Speedup         |
|----------------|----------|-----------------------|-----------------|
| First init     | 3.426s   | 3.546s                | 0.966x (slower) |
| Already init'd | 2.623s   | 1.774s                | 1.479x (faster) |

```python
import signac

project = signac.init_project('test')

for i in range(20000):
    project.open_job({'a': i, 'b': i*2, 'c': i*3}).init()
```

## Types of Changes
- [x] New feature

## Checklist:
- [x] I am familiar with the [**Contributing Guidelines**](https://github.com/glotzerlab/signac/blob/master/CONTRIBUTING.md).
- [x] I agree with the terms of the [**Contributor Agreement**](https://github.com/glotzerlab/signac/blob/master/ContributorAgreement.md).
- [x] My name is on the [list of contributors](https://github.com/glotzerlab/signac/blob/master/contributors.yaml).
- [x] My code follows the [code style guideline](https://github.com/glotzerlab/signac/blob/master/CONTRIBUTING.md#code-style) of this project.
- [x] The changes introduced by this pull request are covered by existing or newly introduced tests.

If necessary:
- [x] I have updated the [changelog](https://github.com/glotzerlab/signac/blob/master/changelog.txt) and added all related issue and pull request numbers for future reference (if applicable). See example below.